+++ components/Icons.tsx
@@ -102,8 +102,8 @@ export function SecurityIcon(props: SVGAttributes<SVGElement>) {
   return (
     <svg fill="currentColor" width="16" height="17" viewBox="0 0 16 17" xmlns="http://www.w3.org/2000/svg" {...props}>
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M7.661 0.122295C7.75629 0.0432577 7.8762 0 8 0C8.1238 0 8.24371 0.0432577 8.339 0.122295C10.3366 1.78395 12.8215 2.74905 15.417 2.87129C15.534 2.87489 15.6459 2.9194 15.7335 2.99705C15.821 3.07471 15.8785 3.1806 15.896 3.29629C15.965 3.81629 16 4.34629 16 4.88629C16 10.0483 12.74 14.4493 8.166 16.1423C8.05886 16.1818 7.94114 16.1818 7.834 16.1423C3.26 14.4493 0 10.0483 0 4.88529C0 4.34729 0.0350001 3.81629 0.104 3.29629C0.121528 3.18044 0.17919 3.07442 0.266919 2.99674C0.354649 2.91906 0.466873 2.87466 0.584 2.87129C3.17927 2.74855 5.66379 1.78411 7.661 0.122295ZM11.857 6.07629C11.974 5.91531 12.0222 5.71445 11.991 5.51791C11.9599 5.32138 11.852 5.14526 11.691 5.02829C11.53 4.91133 11.3292 4.86312 11.1326 4.89425C10.9361 4.92538 10.76 5.03331 10.643 5.19429L7.16 9.98429L5.28 8.10429C5.21078 8.0327 5.128 7.9756 5.03647 7.93634C4.94495 7.89707 4.84653 7.87643 4.74694 7.87561C4.64736 7.87479 4.54861 7.89382 4.45646 7.93157C4.3643 7.96933 4.28059 8.02505 4.2102 8.09551C4.13982 8.16596 4.08417 8.24972 4.0465 8.34191C4.00883 8.4341 3.9899 8.53287 3.99081 8.63245C3.99173 8.73204 4.01246 8.83044 4.05181 8.92193C4.09116 9.01341 4.14834 9.09614 4.22 9.16529L6.72 11.6653C6.79663 11.742 6.88896 11.8011 6.99065 11.8387C7.09233 11.8762 7.20094 11.8913 7.30901 11.8828C7.41708 11.8744 7.52203 11.8426 7.61663 11.7897C7.71123 11.7368 7.79324 11.664 7.857 11.5763L11.857 6.07629Z"
         fill="white"
         fillOpacity="0.5"
