+++ pages/memory-list/index.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled'
 
 import Card from '@/components/Card'
 import Modal from '@/components/Modal'
-import MemoryDetail from '@/components/MemoryDetail'
+import MemoryDetail from '@/components/Modal/MemoryDetail'
 import useControlModal from '@/lib/hooks/useControlModal'
 import createdAtToTitleDate from '@/lib/utils/createdAtToTitleDate'
 import makeYouTubeVideoId from '@/lib/utils/makeYouTubeVideoId'
