+++ pages/upload/index.tsx
@@ -14,7 +14,7 @@ import makeYouTubeVideoId from '@/lib/utils/makeYouTubeVideoId'
 import KeywordFiled from '@/components/KeywordFiled'
 import useCreateSuggestionImage from '@/lib/hooks/useCreateSuggestionImage'
 import axios from 'axios'
-import YouTubePlayer from '@/components/YouTubePlayer'
+import YouTubePlayer from '@/components/button/YouTubePlayerButton'
 import { SubmitHandler } from 'react-hook-form'
 import { UploadFormProps } from '@/lib/types/uploadFormProps'
 import Image from 'next/image'
